CornScanner/requirements.txt
Django==4.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
python-dotenv==1.0.0